requirements.txt
# Core Dependencies - OpenRouter compatible
openai>=1.3.0
python-dotenv>=1.0.0
requests>=2.31.0
aiohttp>=3.9.0

# Date/Time
python-dateutil>=2.8.2

# Scheduling
schedule>=1.2.0

# Data Processing
pandas>=2.1.0
numpy>=1.24.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0

# Development
black>=23.10.0
flake8>=6.1.0

# Flight & Hotel APIs
amadeus>=9.0.0

# Email notifications
email-validator>=2.0.0

# Calendar export
icalendar>=5.0.0
pytz>=2024.1
